download 3rd party just once

+++ 3rdparty/get.sh
@@ -1,2 +1,6 @@
 #!/bin/sh
+if [ -e .apps_downloaded ]; then
+    exit 0
+fi
 wget 'https://rufus.akeo.ie/downloads/rufus-2.11p.exe' -O rufus.exe
+touch .apps_downloaded
